Added window resizing + several algorithms.
Algorithms added:
- C++ version of Andrey Astrelin's GrailSort
- American Flag Sort
- MSD Radix Sort
- LSD Radix Sort
- Bottom-Up Merge Sort using std::inplace_merge

+++ avk/algorithms/insertion_sort.cpp
@@ -2,10 +2,10 @@
 
 void insertion_sort(main_array& array)
 {
-	for (int i = 1; i < (int)array.size(); ++i)
+	for (uint i = 1; i < array.size(); ++i)
 	{
 		item tmp = array[i];
-		int j = i - 1;
+		uint j = i - 1;
 		for (; j >= 0 && array[j] > tmp; --j)
 			array[j + 1] = array[j];
 		array[j + 1] = tmp;
